refactor(auth): split AuthService into Auth/Token/OAuth services
- TokenService (165 lines): issueTokens, refresh with rotation/reuse
  detection, logout
- OAuthService (326 lines): provider login, account linking/unlinking,
  one-time exchange codes, unique username generation
- AuthService slims from 554 to 94 lines: register, loginValidatedUser,
  changePassword — delegates token issuing to TokenService
- Remove deprecated AuthService.login() (no production callers)
- Controller injects all three services; module exports TokenService
  and OAuthService alongside AuthService
- Specs split along the same seams (auth/token/oauth + 3 property
  specs renamed); test helpers build a real TokenService over existing
  mocks so JWT assertions stay intact

Verification: 284/284 suites, 3935/3935 tests; contracts:update diff
empty; live smoke register→login→me→refresh OK.
